completes initial upgrade to webpack 2.1.0-beta.24 & react hot loader 3.0.0-beta.4
# Why
* Because webpack 2 is awesome, provides better validation, cleaner
separation of concerns, native support for ES6, dynamic code splitting,
and a number of other things.
* Refactored to use react-hot-loader 3 since webpack 2 has more natural
support for HMR via event emitters.

# What
* Moved universal, server, and client dirs under src so it’s easier to
set the context in webpack config.
* [WIP] refactor of path resolution since paths may no longer be
functions  and must be a string.
* Moves invalid loader options into LoaderOptions plugin
* Uses the correct, new syntax for referencing es2015 loose preset in
.babelrc

# Side Effects
* Styles are currently broken until postcss-loader patches their config
options to work within the loader query object in the webpack config
(It’s invalid to pass a function to LoaderOptions plugin and invalid to
use postcss in the base webpack config)