migrates to React Router 4
# Why
Because RR4 is about to be released - one more set of breaking changes
(locked the version here to stable)  - and it dramatically simplifies
routing, is much more powerful, allows routes to be components, and is
generally much more pleasant to work with than RR3.

# What
* Simplifies webpack config by converting it from a function to a
simple object - next step is to figure out how to use the webpack cli
to bundle the assets (still doesn’t support ES6 in the config files
because of node)
* Converts `routes` to `router`
* Refactors route config now located at `router/routes`
* Moves Root into `client`
* Adds `router/matches` directory for holding match helpers
* Refactors `Counter` to accept `store` as a prop for dynamic reducer
injection
* The router directory structure now mirrors the router shape, with
each folder containing its child routes. This will allow for better
code-splitting on the route level.