adds session module and boilerplate UI
# Why
So we can have a standardized approach towards authentication and
session management amongst the booster kits.

# What
* Replaces `isomorphic-fetch` with `axis` for client-side HTTP
* Adds `session` module with a common reducer and action directories
for `logIn`, `signUp`, `signOut`, and `resetPassword`.
* Adds `universal/feature` directory to hold completely standalone
features that can be loaded up at any route - with matching features in
the `session` module
* Refactors `lib` to incude `api` and a few other things like test and
error helpers.
* Adds some basic form components
* Adds local storage middleware to the redux store
* Adds a few placeholders for global styles and adds resolve alias to
them in the webpack config
* Adds config and global env for selection the server type (phoenix or
rails)…which are wired up to the corresponding services endpoints
configured in `universal/lib/api`

# Clean Up
* I started writing tests for the session module, but upon further
investigation of trying to make them fail purpose without being able to
do so, I decided to hold off until I could pair with someone to
investigate why we’re getting false positives when we shouldn’t be.