completes full graphql query support
# Why
* The mocking is not working correctly.
* Tests are failing.
* Typings are off.

# What
* Adds a new `mockApollo` Provider that is a complete stand-in
replacement for `ApolloProvider` and `client`. It more or less creates
a mock networkInterface for Apollo, then uses a schema generated from
an introspection query against the JSON schema generated from the
server in order to provide Apollo with an executable schema it can
auto-mock against types (or use custom mocks passed in). It also has
support for passing in initial state, else it defers to the state
derived from the reducers.
* Fixes bug with test execution - `apolli-client` has an optional
dependency specified for react-dom 16.x. Removing it from `yarn.lock`
and reinstalling the packages solved the problem. [Open
Issue](https://github.com/apollographql/react-apollo/issues/826)
* Fixes mock queries for `viewer`.
* Adds better type support for Graphql. Part of this includes removing
the `withData` higher order component for the time being until we can
figure out how to deal with some of the more complex type errors it was
producing. A complete example of the way to wrap a component with a
query can be found in `routes/docs/examples/grapqhl/root.tsx` with a
corresponding test in the same location.
* Adds a `mockGraphql` function for returning a mocked query outside of
a test - like if you want to mock some data out without making changes
to the server.
* Schema generation and type generation from the schema / used queries
is now fully automated. It triggers with `yarn run start` (Dev) and
`yarn run build` (Prod). This prevents the client from ever getting out
of sync with the server and enforces strict type checking on every
query, subscription, or mutation.